+++ mulanbay-pms/src/main/java/cn/mulanbay/pms/web/bean/response/plan/PlanReportPredictVo.java
@@ -7,26 +7,26 @@ public class PlanReportPredictVo extends PlanReport {
     /**
      * 次数预测
      */
-    private Float predictCount;
+    private Double predictCount;
 
     /**
      * 计划值预测
      */
-    private Float predictValue;
+    private Double predictValue;
 
-    public Float getPredictCount() {
+    public Double getPredictCount() {
         return predictCount;
     }
 
-    public void setPredictCount(Float predictCount) {
+    public void setPredictCount(Double predictCount) {
         this.predictCount = predictCount;
     }
 
-    public Float getPredictValue() {
+    public Double getPredictValue() {
         return predictValue;
     }
 
-    public void setPredictValue(Float predictValue) {
+    public void setPredictValue(Double predictValue) {
         this.predictValue = predictValue;
     }
 }
